Move editable grid actions to internal/components/editable/actions.ts
