+++ packages/components/src/internal/components/samples/actions.ts
@@ -18,9 +18,10 @@ import { ActionURL, Ajax, Domain, Filter, Query, Utils } from '@labkey/api';
 
 import { IEntityTypeDetails } from '../entities/models';
 import { deleteEntityType, getSelectedItemSamples } from '../entities/actions';
+import { getSelectedData } from '../editable/actions';
 
 import { Location } from '../../util/URL';
-import { getSelectedData, getSelection, getSnapshotSelections } from '../../actions';
+import { getSelection, getSnapshotSelections } from '../../actions';
 
 import { caseInsensitive, handleRequestFailure } from '../../util/utils';
