+++ packages/components/src/internal/models.ts
@@ -1,15 +1,5 @@
 import { SchemaQuery } from '../public/SchemaQuery';
 
-import { ProductMenuModel } from './components/navigation/model';
-import { AppURL } from './url/AppURL';
-
-export interface CommonPageProps {
-    goBack?: (n?: number) => void;
-    menu: ProductMenuModel;
-    menuInit?: (invalidate?: boolean) => void;
-    navigate?: (url: string | AppURL, replace?: boolean) => void;
-}
-
 export function createGridModelId(gridId: string, schemaQuery: SchemaQuery, keyValue?: any): string {
     const parts = [gridId, schemaQuery.getKey()];
